refactor mobile-client types and signing logic: consolidate signature parameter types, restructure server response types, and enhance type definitions for improved clarity and maintainability across the mobile-client package
