+++ packages/engine-client/src/types/clientExecuteTypes.ts
@@ -9,6 +9,7 @@ import {
   EIP712TransferQuoteParams,
   EIP712WithdrawCollateralParams,
   EIP712WithdrawCollateralV2Params,
+  SignatureParams,
 } from '@nadohq/shared';
 import BigNumber from 'bignumber.js';
 import { EngineServerExecuteSuccessResult } from './serverExecuteTypes';
@@ -17,12 +18,8 @@ import { EngineServerExecuteSuccessResult } from './serverExecuteTypes';
  * Either verifying address or signature must be provided;
  * If signature is not provided, the verifying address with the engine signer will be used to sign.
  */
-export type SignatureParams =
-  | {
-      // Endpoint address for all executes except order placement
-      verifyingAddr: string;
-      chainId: number;
-    }
+export type SignatureParamsOrSignature =
+  | SignatureParams
   | {
       signature: string;
     };
@@ -38,7 +35,7 @@ export type WithSignature<T> = T & {
 };
 
 // Params associated with all engine executes
-export type WithBaseEngineExecuteParams<T> = SignatureParams &
+export type WithBaseEngineExecuteParams<T> = SignatureParamsOrSignature &
   Omit<T, 'nonce'> & {
     nonce?: string;
   };
