+++ packages/mobile-client/src/index.ts
@@ -2,3 +2,4 @@ export * from './endpoints';
 export * from './MobileClient';
 export * from './signing';
 export * from './types';
+export * from './utils';
