+++ packages/mobile-client/src/types/MobileServerFailureError.ts
@@ -1,7 +1,7 @@
 import {
   MobileServerFailureResponse,
   MobileServerRequestType,
-} from './serverTypes';
+} from './serverBaseTypes';
 
 /**
  * Error thrown when the mobile service API returns a failure envelope, either as a non-2xx HTTP response
