+++ packages/mobile-client/src/types/index.ts
@@ -1,4 +1,7 @@
 export * from './clientTypes';
 export * from './mobileErrorCodes';
 export * from './MobileServerFailureError';
-export * from './serverTypes';
+export * from './serverBaseTypes';
+export * from './serverExecuteTypes';
+export * from './serverModelTypes';
+export * from './serverQueryTypes';
