+++ packages/mobile-client/src/types/serverModelTypes.ts
@@ -1,0 +1,74 @@
+import { Hex } from 'viem';
+
+/**
+ * Platform of a device registered for push notifications.
+ */
+export type MobileNotificationPlatform = 'ios' | 'android';
+
+/**
+ * Category of push notification.
+ */
+export type MobileNotificationCategory =
+  | 'order_fill'
+  | 'order_update'
+  | 'liquidation'
+  | 'funding'
+  | 'product_listing'
+  | 'announcement';
+
+/**
+ * Server-side public profile shape (snake_case).
+ */
+export interface MobileServerProfile {
+  subaccount: Hex;
+  username: string;
+  display_name: string;
+}
+
+/**
+ * Server-side identity shape (snake_case) returned by the signed `self_identity` query.
+ */
+export interface MobileServerIdentity {
+  subaccount: Hex;
+  username: string;
+  display_name: string;
+  private_mode: boolean;
+}
+
+/**
+ * Server-side scope limiting a notification category preference. Scopes are rejected by the backend for the
+ * MVP (`scopes` must be empty) but are part of the wire format.
+ */
+export type MobileServerNotificationPreferenceScope =
+  | { type: 'subaccount'; subaccount: Hex }
+  | { type: 'product'; product_id: number };
+
+/**
+ * Server-side per-category notification preference (snake_case).
+ */
+export interface MobileServerNotificationCategoryPreference {
+  category: MobileNotificationCategory;
+  enabled: boolean;
+  scopes: MobileServerNotificationPreferenceScope[];
+}
+
+/**
+ * Server-side notification preferences shape (snake_case), used in both the `update_preferences` execute and
+ * the `notification_preferences` query response.
+ */
+export interface MobileServerNotificationPreferences {
+  /** Only `1` is accepted by the backend for the MVP. */
+  schema_version: 1;
+  categories: MobileServerNotificationCategoryPreference[];
+}
+
+/**
+ * Server-side registered push device shape (snake_case).
+ */
+export interface MobileServerRegisteredDevice {
+  platform: MobileNotificationPlatform;
+  locale: string | null;
+  app_version: string | null;
+  token_fingerprint_prefix: string;
+  last_seen_at: number;
+}
