+++ packages/shared/src/eip712/eip712ValueTypes.ts
@@ -74,12 +74,8 @@ export type EIP712MintNlpValues = WithEIP712Sender<EIP712MintNlpParams>;
 
 export type EIP712BurnNlpValues = WithEIP712Sender<EIP712BurnNlpParams>;
 
-// `sender` is already bytes32-encoded by the caller, so the values mirror the params exactly. Written as a
-// mapped type (not `= EIP712NadoAuthenticationParams`) so it gains an implicit index signature and stays
-// assignable to viem's `signTypedData` message (`Record<string, unknown>`), which interfaces do not.
-export type EIP712NadoAuthenticationValues = {
-  [K in keyof EIP712NadoAuthenticationParams]: EIP712NadoAuthenticationParams[K];
-};
+export type EIP712NadoAuthenticationValues =
+  WithEIP712Sender<EIP712NadoAuthenticationParams>;
 
 /**
  * All possible requests to be signed, to the EIP712 value interface
