+++ packages/trigger-client/src/types/clientTypes.ts
@@ -6,6 +6,7 @@ import {
   EIP712CancelOrdersParams,
   EIP712CancelProductOrdersParams,
   OrderAppendix,
+  SignatureParams,
   Subaccount,
 } from '@nadohq/shared';
 import BigNumber from 'bignumber.js';
@@ -17,15 +18,6 @@ import {
 } from './serverQueryTypes';
 
 type WithOptionalNonce<T> = Omit<T, 'nonce'> & { nonce?: string };
-
-interface SignatureParams {
-  /**
-   * Address derived from productId for placement (see getOrderVerifyingAddr)
-   * endpoint address for cancellation & listing
-   */
-  verifyingAddr: string;
-  chainId: number;
-}
 
 /**
  * Executes
